requirements.txt
python-dotenv>=0.20.0
fastapi>=0.78.0
pymysql>=1.0.2
aiomysql>=0.1.1
colorama>=0.4.5
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
#sqlalchemy[asyncio]>=1.4.39
databases>=0.6.0
python-multipart>=0.0.5
uvicorn>=0.18.1
# blacksheep is annoying (we can't use jinja 3.1.2+ for now)
jinja2==3.0.2